waveform/requirements.txt
numpy >1, <2
pyprecice~=2.0
scipy
jinja2
pandas
